+++ client/python/tests/test_client.py
@@ -41,14 +41,22 @@ def test_client_fails_to_create_with_wrong_url():
     with pytest.raises(ValueError):
         OpenLineageClient(url="196.168.0.1", session=session)
 
+
 def test_client_passes_to_create_with_valid_url():
     session = MagicMock()
-    assert OpenLineageClient(url="http://196.168.0.1", session=session).url == "http://196.168.0.1"
-    assert OpenLineageClient(url="http://196.168.0.1", session=session).url == "http://196.168.0.1"
-    assert OpenLineageClient(url="http://example.com  ", session=session).url == "http://example.com"
-    assert OpenLineageClient(url=" http://example.com", session=session).url == "http://example.com"
-    assert OpenLineageClient(url="  http://marquez:5000  ", session=session).url == "http://marquez:5000"
-    assert OpenLineageClient(url="  https://marquez  ", session=session).url == "https://marquez"
+    assert OpenLineageClient(url="http://196.168.0.1", session=session).url == \
+           "http://196.168.0.1"
+    assert OpenLineageClient(url="http://196.168.0.1", session=session).url == \
+           "http://196.168.0.1"
+    assert OpenLineageClient(url="http://example.com  ", session=session).url == \
+           "http://example.com"
+    assert OpenLineageClient(url=" http://example.com", session=session).url == \
+           "http://example.com"
+    assert OpenLineageClient(url="  http://marquez:5000  ", session=session).url == \
+           "http://marquez:5000"
+    assert OpenLineageClient(url="  https://marquez  ", session=session).url == \
+           "https://marquez"
+
 
 def test_client_sends_proper_json_with_minimal_event():
     session = MagicMock()