+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Added
 * facet: adding custom spark version facet to spark integration [@OleksandrDvornik](https://github.com/OleksandrDvornik)
 
+### Fixed
+
+* strip openlineage url in python client [@OleksandrDvornik](https://github.com/OleksandrDvornik)
+
 ## [0.2.3](https://github.com/OpenLineage/OpenLineage/releases/tag/0.2.3) - 2021-10-07
 
 ### Fixed